# Conflicts:
#	src/Lykke.Service.PayInternal.AzureRepositories/Transaction/PaymentRequestTransactionEntity.cs
#	src/Lykke.Service.PayInternal.Core/Domain/Transfer/ITransferRepository.cs
#	src/Lykke.Service.PayInternal.Services/BtcTransferService.cs
#	src/Lykke.Service.PayInternal/Rabbit/Publishers/PaymentRequestPublisher.cs

+++ src/Lykke.Service.PayInternal.Services/Extensions.cs
@@ -22,7 +22,7 @@ namespace Lykke.Service.PayInternal.Services
             return new CreateTransaction
             {
                 WalletAddress = src.WalletAddress,
-                Amount = src.Amount,
+                Amount = (decimal) src.Amount,
                 FirstSeen = src.FirstSeen,
                 Confirmations = src.Confirmations,
                 BlockId = src.BlockId,
